requirements.txt
numpy
matplotlib
Pillow
tqdm
scikit-learn
pycocotools
transformers
tensorboard
opencv-python
open_clip_torch
einops
accelerate
datasets
peft
bitsandbytes
torchpack
gradio
evaluate
scikit-image
diffusers
# flash-attn # only for A100/A800
opencv-python
tqdm
matplotlib
scipy
segment-anything